requirements.txt
requests>=2.31.0
pandas>=2.0.0
streamlit>=1.30.0
plotly>=5.18.0
gspread>=6.0.0
oauth2client>=4.1.3
python-dotenv>=1.0.0
apscheduler>=3.10.0
